requirements.txt
requests
Django==2.2.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
python-magic==0.4.15
requests==2.22.0
